python-app/requirements.txt
attrs==25.3.0
azure-core==1.34.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==43.0.3
dnspython==2.7.0
exceptiongroup==1.3.0
greenlet==3.2.2
h11==0.16.0
idna==3.10
isodate==0.7.2
lxml==5.4.0
msal==1.32.3
msal-extensions==1.3.1
outcome==1.3.0.post0
packaging==25.0
playwright==1.52.0
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.23.0
pyee==13.0.0
PyJWT==2.10.1
pymongo==4.13.0
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
streamlink @ git+https://github.com/streamlink/streamlink.git@cd27285adbf1a5248bc563420e38477f7159c732
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
